refactor: correct firebase util file names
