+++ src/__tests__/app/api/auth/invalidateLogin.test.tsx
@@ -10,7 +10,7 @@ const mocks = vi.hoisted(() => {
   };
 });
 
-vi.mock('@/lib/firebase/firebase-admin-config', () => {
+vi.mock('@/lib/firebase/firebaseAdminConfig', () => {
   return { adminAuth: mocks.adminAuth() };
 });
