+++ src/app/components/auth/signin/Login.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { getRedirectResult, signInWithRedirect } from 'firebase/auth';
 import { useContext, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
-import { auth, provider } from '@/lib/firebase/firebase-config';
+import { auth, provider } from '@/lib/firebase/firebaseConfig';
 import { LoginForm } from '@/app/components/auth/signin/LoginForm';
 import { locale } from '@/locales/locale';
 import { localeContext } from '@/locales/localeProvider';
